refactoring epirus new/import code
